+++ apps/prairielearn/src/server.ts
@@ -15,7 +15,6 @@ import * as url from 'url';
 import blocked from 'blocked';
 import blockedAt from 'blocked-at';
 import bodyParser from 'body-parser';
-import Bowser from 'bowser';
 import cookieParser from 'cookie-parser';
 import esMain from 'es-main';
 import express, {
@@ -26,7 +25,6 @@ import express, {
   type Response,
 } from 'express';
 import asyncHandler from 'express-async-handler';
-import _ from 'lodash';
 import multer from 'multer';
 import onFinished from 'on-finished';
 import passport from 'passport';
@@ -176,15 +174,6 @@ export async function initExpress(): Promise<Express> {
 
   // API routes don't utilize sessions; don't run the session/flash middleware for them.
   app.use(excludeRoutes(['/pl/api'], sessionRouter));
-
-  app.use(function (req, res, next) {
-    if (req.headers['user-agent']) {
-      res.locals.userAgent = Bowser.parse(req.headers['user-agent']);
-    } else {
-      res.locals.userAgent = null;
-    }
-    next();
-  });
 
   // special parsing of file upload paths -- this is inelegant having it
   // separate from the route handlers but it seems to be necessary
@@ -921,6 +910,10 @@ export async function initExpress(): Promise<Express> {
         )
       ).default,
     ],
+  );
+  app.use(
+    '/pl/course_instance/:course_instance_id(\\d+)/instructor/assessment/:assessment_id(\\d+)/manual_grading/assessment_question/:assessment_question_id(\\d+)/ai_grading_runs',
+    (await import('./ee/pages/instructorAiGradingRuns/instructorAiGradingRuns.js')).default,
   );
   app.use(
     '/pl/course_instance/:course_instance_id(\\d+)/instructor/assessment/:assessment_id(\\d+)/manual_grading/instance_question/:instance_question_id(\\d+)',
@@ -1737,6 +1730,13 @@ export async function initExpress(): Promise<Express> {
       coreElements: false,
     }),
   );
+  app.use(/^(\/pl\/public\/course_instance\/[0-9]+\/assessment\/[0-9]+)\/?$/, (req, res, _next) => {
+    res.redirect(`${req.params[0]}/questions`);
+  });
+  app.use(
+    '/pl/public/course_instance/:course_instance_id(\\d+)/assessment/:assessment_id(\\d+)/questions',
+    (await import('./pages/publicAssessmentQuestions/publicAssessmentQuestions.js')).default,
+  );
 
   // Client files for questions
   app.use(
@@ -1886,7 +1886,9 @@ export async function initExpress(): Promise<Express> {
   app.use((err: any, req: Request, res: Response, next: NextFunction) => {
     const chars = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ'.split('');
 
-    res.locals.error_id = _.times(12, () => _.sample(chars)).join('');
+    res.locals.error_id = Array.from({ length: 12 })
+      .map(() => chars[Math.floor(Math.random() * chars.length)])
+      .join('');
 
     err.status = err.status ?? maybeGetStatusCodeFromSqlError(err) ?? 500;
 
@@ -2082,23 +2084,6 @@ if (esMain(import.meta) && config.startServer) {
         // enabled. Otherwise, allow Sentry to install its own stuff so
         // that request isolation works correctly.
         skipOpenTelemetrySetup: config.openTelemetryEnabled,
-
-        beforeSend: (event) => {
-          // This will be necessary until we can consume the following change:
-          // https://github.com/chimurai/http-proxy-middleware/pull/823
-          //
-          // The following error message should match the error that's thrown
-          // from the `router` function in our `http-proxy-middleware` config.
-          if (
-            event.exception?.values?.some(
-              (value) => value.type === 'Error' && value.value === 'Workspace is not running',
-            )
-          ) {
-            return null;
-          }
-
-          return event;
-        },
       });
     }