+++ packages/html-ejs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/html-ejs
 
+## 2.0.13
+
+### Patch Changes
+
+- 82f9c2f: Upgrade all JavaScript dependencies
+- Updated dependencies [82f9c2f]
+  - @prairielearn/html@4.0.13
+
+## 2.0.12
+
+### Patch Changes
+
+- 03f1008: Upgrade all JavaScript dependencies
+- Updated dependencies [03f1008]
+  - @prairielearn/html@4.0.12
+
 ## 2.0.11
 
 ### Patch Changes