+++ packages/node-metrics/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/node-metrics
 
+## 1.0.3
+
+### Patch Changes
+
+- 82f9c2f: Upgrade all JavaScript dependencies
+
+## 1.0.2
+
+### Patch Changes
+
+- 03f1008: Upgrade all JavaScript dependencies
+
 ## 1.0.1
 
 ### Patch Changes