+++ packages/opentelemetry/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/opentelemetry
 
+## 2.1.13
+
+### Patch Changes
+
+- 82f9c2f: Upgrade all JavaScript dependencies
+
+## 2.1.12
+
+### Patch Changes
+
+- 03f1008: Upgrade all JavaScript dependencies
+
 ## 2.1.11
 
 ### Patch Changes