+++ packages/postgres/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @prairielearn/postgres
 
+## 2.1.11
+
+### Patch Changes
+
+- 82f9c2f: Upgrade all JavaScript dependencies
+
+## 2.1.10
+
+### Patch Changes
+
+- c24120e: Minor changes to reduce dependency on lodash
+- 03f1008: Upgrade all JavaScript dependencies
+
 ## 2.1.9
 
 ### Patch Changes